:recycle: Chore: Clean up @tevm/procedures
