+++ .changeset/good-cups-report.md
@@ -1,0 +1,5 @@
+---
+"@tevm/procedures": patch
+---
+
+Refactored tevm/procedures and found small bugs along the way
