+++ packages/procedures/src/anvil/anvilDumpStateProcedure.js
@@ -1,0 +1,22 @@
+import { dumpStateProcedure } from '../dumpstate/dumpStateProcedure.js'
+
+// TODO this procedure is implemented completely wrong. Anvil encodes the state into a single hex string? Or maybe the type is wrong?
+/**
+ * @experimental
+ * Request handler for anvil_dumpState JSON-RPC requests.
+ * @param {import('@tevm/base-client').BaseClient} client
+ * @returns {import('./AnvilProcedure.js').AnvilDumpStateProcedure}
+ */
+export const anvilDumpStateJsonRpcProcedure = (client) => {
+	console.warn('anvilDumpStateJsonRpcProcedure is not implemented correctly')
+	return async (request) => {
+		return /** @type any*/ ({
+			...(await dumpStateProcedure(client)({
+				...(request.id ? { id: request.id } : {}),
+				jsonrpc: '2.0',
+				method: 'tevm_dumpState',
+			})),
+			method: request.method,
+		})
+	}
+}
