+++ packages/procedures/src/requestProcedure.js
@@ -1,52 +1,5 @@
-import {
-	chainIdHandler,
-	ethGetLogsHandler,
-	ethSendTransactionHandler,
-	forkAndCacheBlock,
-	testAccounts,
-	traceCallHandler,
-} from '@tevm/actions'
-import { Block, BlockHeader } from '@tevm/block'
-import { InvalidParamsError, MethodNotFoundError, MethodNotSupportedError } from '@tevm/errors'
-import { createJsonRpcFetcher } from '@tevm/jsonrpc'
-import { BlobEIP4844Transaction, TransactionFactory } from '@tevm/tx'
-import {
-	EthjsAccount,
-	EthjsAddress,
-	bytesToHex,
-	getAddress,
-	hexToBigInt,
-	hexToBytes,
-	hexToNumber,
-	numberToHex,
-} from '@tevm/utils'
-import { runTx } from '@tevm/vm'
-import { version as packageJsonVersion } from '../package.json'
-import { ethAccountsProcedure } from './eth/ethAccountsProcedure.js'
-import { ethCallProcedure } from './eth/ethCallProcedure.js'
-import { ethGetTransactionReceiptJsonRpcProcedure } from './eth/ethGetTransactionReceiptProcedure.js'
-import { ethSignProcedure } from './eth/ethSignProcedure.js'
-import { ethSignTransactionProcedure } from './eth/ethSignTransactionProcedure.js'
-import {
-	blockNumberProcedure,
-	callProcedure,
-	chainIdProcedure,
-	dumpStateProcedure,
-	ethGetLogsProcedure,
-	gasPriceProcedure,
-	getAccountProcedure,
-	getBalanceProcedure,
-	getCodeProcedure,
-	getStorageAtProcedure,
-	loadStateProcedure,
-	mineProcedure,
-	scriptProcedure,
-	setAccountProcedure,
-} from './index.js'
-import { blockToJsonRpcBlock } from './utils/blockToJsonRpcBlock.js'
-import { generateRandomId } from './utils/generateRandomId.js'
-import { parseBlockTag } from './utils/parseBlockTag.js'
-import { txToJsonRpcTx } from './utils/txToJsonRpcTx.js'
+import { MethodNotFoundError } from '@tevm/errors'
+import { createHandlers } from './createHandlers.js'
 
 /**
  * Request handler for JSON-RPC requests.
@@ -77,1248 +30,22 @@ import { txToJsonRpcTx } from './utils/txToJsonRpcTx.js'
  * ```
  */
 export const requestProcedure = (client) => {
+	const allHandlers = createHandlers(client)
 	return async (request) => {
 		await client.ready()
 		client.logger.debug(request, 'JSON-RPC request received')
-		switch (request.method) {
-			case 'tevm_call':
-				return /**@type any*/ (callProcedure(client)(request))
-			case /** @type {any} */ ('tevm_contract'): {
-				const err = new MethodNotSupportedError(
-					'UnsupportedMethodError: tevm_contract is not supported. Encode the contract arguments and use tevm_call instead.',
-				)
-				return /**@type any*/ ({
-					id: /** @type any*/ (request).id,
-					jsonrpc: '2.0',
-					error: {
-						code: err._tag,
-						message: err.message,
-					},
-				})
-			}
-			case 'tevm_getAccount':
-				return /**@type any*/ (getAccountProcedure)(client)(request)
-			case 'tevm_setAccount':
-				return /**@type any*/ (setAccountProcedure)(client)(request)
-			case 'tevm_script':
-				return /**@type any*/ (scriptProcedure)(client)(request)
-			case 'eth_blockNumber':
-				return /** @type any */ (blockNumberProcedure(client)(request))
-			case 'tevm_dumpState':
-				return /** @type any */ (dumpStateProcedure)(client)(request)
-			case 'tevm_loadState': {
-				return /** @type any */ (loadStateProcedure)(client)(request)
-			}
-			case 'eth_chainId':
-				return /** @type any */ (chainIdProcedure(client)(request))
-			case 'eth_call':
-				return /** @type any */ (ethCallProcedure(client)(request))
-			case 'eth_getCode':
-				return /** @type any */ (getCodeProcedure(client)(request))
-			case 'eth_getStorageAt':
-				return /** @type any */ (getStorageAtProcedure(client)(request))
-			case 'eth_gasPrice':
-				// TODO this vm.blockchain should not be type any
-				return /** @type any */ (gasPriceProcedure(client)(request))
-			case 'eth_getBalance':
-				return /** @type any */ (getBalanceProcedure(client)(request))
-			// TODO move all wallet methods to seperate decorator
-			case 'eth_sign':
-				return ethSignProcedure(testAccounts)(request)
-			case 'eth_signTransaction':
-				return ethSignTransactionProcedure({
-					accounts: testAccounts,
-					getChainId: () => chainIdHandler(client)().then((bigNumber) => Number(bigNumber)),
-				})(request)
-			case 'eth_accounts':
-				return ethAccountsProcedure(testAccounts)(request)
-			case 'eth_mining':
-				return Promise.resolve(false)
-			case 'eth_syncing':
-				return Promise.resolve(false)
-			case 'anvil_setCode':
-			case /** @type {'anvil_setCode'}*/ ('ganache_setCode'):
-			case /** @type {'anvil_setCode'}*/ ('hardhat_setCode'): {
-				/**
-				 * @type {import('./anvil/AnvilJsonRpcRequest.js').AnvilSetCodeJsonRpcRequest}
-				 */
-				const codeRequest = request
-				const result = setAccountProcedure(client)({
-					jsonrpc: codeRequest.jsonrpc,
-					method: 'tevm_setAccount',
-					params: [{ address: codeRequest.params[0], deployedBytecode: codeRequest.params[1] }],
-					...(codeRequest.id ? { id: codeRequest.id } : {}),
-				})
-				return {
-					...result,
-					method: codeRequest.method,
-				}
-			}
-			case 'anvil_setBalance':
-			case /** @type {'anvil_setBalance'}*/ ('ganache_setBalance'):
-			case /** @type {'anvil_setBalance'}*/ ('hardhat_setBalance'): {
-				/**
-				 * @type {import('./anvil/AnvilJsonRpcRequest.js').AnvilSetBalanceJsonRpcRequest}
-				 */
-				const balanceRequest = request
-				const balanceResult = setAccountProcedure(client)({
-					jsonrpc: balanceRequest.jsonrpc,
-					method: 'tevm_setAccount',
-					params: [
-						{
-							address: balanceRequest.params[0],
-							balance: balanceRequest.params[1],
-						},
-					],
-					...(balanceRequest.id ? { id: balanceRequest.id } : {}),
-				})
-				return {
-					...balanceResult,
-					method: balanceRequest.method,
-				}
-			}
-			case 'anvil_setNonce':
-			case /** @type {'anvil_setNonce'}*/ ('ganache_setNonce'):
-			case /** @type {'anvil_setNonce'}*/ ('hardhat_setNonce'): {
-				/**
-				 * @type {import('./anvil/AnvilJsonRpcRequest.js').AnvilSetNonceJsonRpcRequest}
-				 **/
-				const nonceRequest = request
-				const nonceResult = setAccountProcedure(client)({
-					jsonrpc: nonceRequest.jsonrpc,
-					method: 'tevm_setAccount',
-					params: [
-						{
-							address: nonceRequest.params[0],
-							nonce: nonceRequest.params[1],
-						},
-					],
-					...(nonceRequest.id ? { id: nonceRequest.id } : {}),
-				})
-				return {
-					...nonceResult,
-					method: nonceRequest.method,
-				}
-			}
-			case 'anvil_setChainId':
-			case /** @type {'anvil_setChainId'}*/ ('hardhat_setChainId'):
-			case /** @type {'anvil_setChainId'}*/ ('ganache_setChainId'): {
-				const chainId = /** @type {import('./anvil/index.ts').AnvilSetChainIdJsonRpcRequest} */ (request).params[0]
-				if (!Number.isInteger(chainId) || hexToNumber(chainId) <= 0) {
-					return {
-						...(request.id ? { id: request.id } : {}),
-						method: request.method,
-						jsonrpc: request.jsonrpc,
-						error: {
-							code: -32602,
-							message: `Invalid id ${chainId}. Must be a positive integer.`,
-						},
-					}
-				}
-				client.logger.warn('Warning: setChainId is currently a noop')
-				return {
-					...(request.id ? { id: request.id } : {}),
-					method: request.method,
-					jsonrpc: request.jsonrpc,
-					result: true,
-				}
-			}
-			case 'eth_coinbase': {
-				if (client.mode === 'normal') {
-					return {
-						...(request.id ? { id: request.id } : {}),
-						method: request.method,
-						jsonrpc: request.jsonrpc,
-						// same default as hardhat
-						result: await client
-							.getVm()
-							.then((vm) => vm.blockchain.getCanonicalHeadBlock())
-							.then((block) => block.header.coinbase),
-					}
-				}
-				if (!client.forkTransport) {
-					throw new Error(
-						'Fatal error! Client is in mode fork or proxy but no forkUrl is set! This indicates a bug in the client.',
-					)
-				}
-				const fetcher = createJsonRpcFetcher(client.forkTransport)
-				return fetcher.request(/** @type any*/ (request))
-			}
-			// TODO move this to it's own procedure
-			case 'eth_sendTransaction': {
-				const sendTransactionRequest = /** @type {import('./eth/index.js').EthSendTransactionJsonRpcRequest}*/ (request)
-				const txHash = await ethSendTransactionHandler(client)({
-					from: request.params[0].from,
-					...(request.params[0].data ? { data: request.params[0].data } : {}),
-					...(request.params[0].to ? { to: request.params[0].to } : {}),
-					...(request.params[0].gas ? { gas: hexToBigInt(request.params[0].gas) } : {}),
-					...(request.params[0].gasPrice ? { gasPrice: hexToBigInt(request.params[0].gasPrice) } : {}),
-					...(request.params[0].value ? { value: hexToBigInt(request.params[0].value) } : {}),
-				})
-				return {
-					method: sendTransactionRequest.method,
-					result: txHash,
-					jsonrpc: '2.0',
-					...(sendTransactionRequest.id ? { id: sendTransactionRequest.id } : {}),
-				}
-			}
-			// TODO move this to it's own procedure
-			case 'eth_sendRawTransaction': {
-				const sendRawTransactionRequest = /** @type {import('./eth/index.js').EthSendRawTransactionJsonRpcRequest}*/ (
-					request
-				)
-				const vm = await client.getVm()
-				const [serializedTx] = request.params
-				const txBuf = hexToBytes(serializedTx)
-				// Blob Transactions sent over RPC are expected to be in Network Wrapper format
-				const tx =
-					txBuf[0] === 0x03
-						? BlobEIP4844Transaction.fromSerializedBlobTxNetworkWrapper(txBuf, { common: vm.common.ethjsCommon })
-						: TransactionFactory.fromSerializedData(txBuf, { common: vm.common.ethjsCommon })
-				if (!tx.isSigned()) {
-					const err = new InvalidParamsError('Transaction must be signed!')
-					return {
-						method: sendRawTransactionRequest.method,
-						jsonrpc: '2.0',
-						...(sendRawTransactionRequest.id ? { id: sendRawTransactionRequest.id } : {}),
-						error: {
-							code: err._tag,
-							message: err.message,
-						},
-					}
-				}
-				const txPool = await client.getTxPool()
-				await txPool.add(tx, true)
-				return {
-					method: sendRawTransactionRequest.method,
-					result: bytesToHex(tx.hash()),
-					jsonrpc: '2.0',
-					...(sendRawTransactionRequest.id ? { id: sendRawTransactionRequest.id } : {}),
-				}
-			}
-			// TODO move this to it's own procedure
-			case 'eth_estimateGas': {
-				const estimateGasRequest = /** @type {import('./eth/EthJsonRpcRequest.js').EthEstimateGasJsonRpcRequest}*/ (
-					request
-				)
-				const callResult = await callProcedure(client)({
-					...estimateGasRequest,
-					params: [...estimateGasRequest.params],
-					method: 'tevm_call',
-				})
-				if (callResult.error || !callResult.result) {
-					return {
-						...callResult,
-						method: estimateGasRequest.method,
-					}
-				}
-				return {
-					method: estimateGasRequest.method,
-					result: callResult.result.totalGasSpent ?? callResult.result.executionGasUsed,
-					jsonrpc: '2.0',
-					...(estimateGasRequest.id ? { id: estimateGasRequest.id } : {}),
-				}
-			}
-			case 'anvil_getAutomine':
-			case /** @type {'anvil_getAutomine'}*/ ('hardhat_getAutomine'):
-			case /** @type {'anvil_getAutomine'}*/ ('ganache_getAutomine'):
-				return {
-					jsonrpc: '2.0',
-					method: request.method,
-					result: client.miningConfig.type === 'auto',
-					...(request.id ? { id: request.id } : {}),
-				}
-			case 'anvil_setCoinbase':
-			case /** @type {'anvil_setCoinbase'}*/ ('hardhat_setCoinbase'):
-			case /** @type {'anvil_setCoinbase'}*/ ('ganache_setCoinbase'): {
-				const vm = await client.getVm()
-				const currentBlock = await vm.blockchain.getCanonicalHeadBlock()
-				const newHeader = BlockHeader.fromHeaderData(
-					{
-						...currentBlock.header.raw(),
-						coinbase: request.params[0],
-					},
-					{
-						common: vm.common,
-						freeze: false,
-						setHardfork: false,
-					},
-				)
-				// TODO this as any is not necessary we shouldn't be doing this instead fix types please
-				const newBlock = Block.fromBlockData(
-					/** @type {any}*/ ({
-						...currentBlock,
-						withdrawals: currentBlock.withdrawals,
-						header: newHeader,
-					}),
-					{
-						common: vm.common,
-						freeze: false,
-						setHardfork: false,
-					},
-				)
-				vm.blockchain.putBlock(newBlock)
-				return {
-					method: request.method,
-					result: true,
-					jsonrpc: '2.0',
-					...(request.id ? { id: request.id } : {}),
-				}
-			}
-			case 'anvil_mine':
-			case 'tevm_mine': {
-				const res = await /** @type any */ (mineProcedure)(client)(request)
-				return {
-					...res,
-					method: request.method,
-				}
-			}
-			case 'debug_traceCall': {
-				const debugTraceCallRequest =
-					/** @type {import('./debug/index.js').DebugTraceCallJsonRpcRequest}*/
-					(request)
-				const { blockTag, tracer, to, gas, data, from, value, timeout, gasPrice, tracerConfig } =
-					debugTraceCallRequest.params[0]
-				const traceResult = await traceCallHandler(client)({
-					tracer,
-					...(to !== undefined ? { to } : {}),
-					...(from !== undefined ? { from } : {}),
-					...(gas !== undefined ? { gas: hexToBigInt(gas) } : {}),
-					...(gasPrice !== undefined ? { gasPrice: hexToBigInt(gasPrice) } : {}),
-					...(value !== undefined ? { value: hexToBigInt(value) } : {}),
-					...(data !== undefined ? { data } : {}),
-					...(blockTag !== undefined ? { blockTag } : {}),
-					...(timeout !== undefined ? { timeout } : {}),
-					...(tracerConfig !== undefined ? { tracerConfig } : {}),
-				})
-				return {
-					method: debugTraceCallRequest.method,
-					result: {
-						gas: numberToHex(traceResult.gas),
-						failed: traceResult.failed,
-						returnValue: traceResult.returnValue,
-						structLogs: traceResult.structLogs.map((log) => {
-							return {
-								gas: numberToHex(log.gas),
-								gasCost: numberToHex(log.gasCost),
-								op: log.op,
-								pc: log.pc,
-								stack: log.stack,
-								depth: log.depth,
-							}
-						}),
-					},
-					jsonrpc: '2.0',
-					...(debugTraceCallRequest.id ? { id: debugTraceCallRequest.id } : {}),
-				}
-			}
-			case 'eth_getTransactionReceipt': {
-				const getTransactionReceiptRequest =
-					/** @type {import('./eth/EthJsonRpcRequest.js').EthGetTransactionReceiptJsonRpcRequest}*/
-					(request)
-				return ethGetTransactionReceiptJsonRpcProcedure(client)(getTransactionReceiptRequest)
-			}
-			case 'eth_getLogs': {
-				const ethGetLogsRequest =
-					/** @type {import('./eth/EthJsonRpcRequest.js').EthGetLogsJsonRpcRequest}*/
-					(request)
-				return ethGetLogsProcedure(client)(ethGetLogsRequest)
-			}
-			case 'eth_getBlockByHash': {
-				const getBlockByHashRequest =
-					/** @type {import('./eth/EthJsonRpcRequest.js').EthGetBlockByHashJsonRpcRequest}*/
-					(request)
-				const vm = await client.getVm()
-				const block = await vm.blockchain.getBlock(hexToBytes(request.params[0]))
-				const includeTransactions = getBlockByHashRequest.params[1] ?? false
-				const result = blockToJsonRpcBlock(block, includeTransactions)
-				return {
-					method: getBlockByHashRequest.method,
-					result,
-					jsonrpc: '2.0',
-					...(getBlockByHashRequest.id ? { id: getBlockByHashRequest.id } : {}),
-				}
-			}
-			case 'eth_getBlockByNumber': {
-				const getBlockByHashRequest =
-					/** @type {import('./eth/EthJsonRpcRequest.js').EthGetBlockByNumberJsonRpcRequest}*/
-					(request)
-				const vm = await client.getVm()
-				const blockTagOrNumber = getBlockByHashRequest.params[0]
-				const block = await (() => {
-					if (blockTagOrNumber.startsWith('0x')) {
-						return vm.blockchain.getBlock(hexToBigInt(/** @type {import('@tevm/utils').Hex}*/ (blockTagOrNumber)))
-					}
-					return vm.blockchain.blocksByTag.get(/** @type {import('@tevm/utils').BlockTag}*/ (blockTagOrNumber))
-				})()
-
-				if (!block && client.forkTransport) {
-					const fetcher = createJsonRpcFetcher(client.forkTransport)
-					return fetcher.request({
-						jsonrpc: '2.0',
-						id: request.id ?? 1,
-						method: 'eth_getBlockByNumber',
-						params: [blockTagOrNumber, getBlockByHashRequest.params[1] ?? false],
-					})
-				}
-				if (!block) {
-					return {
-						...(request.id ? { id: request.id } : {}),
-						method: request.method,
-						jsonrpc: request.jsonrpc,
-						error: {
-							code: -32602,
-							message: `Invalid block tag ${blockTagOrNumber}`,
-						},
-					}
-				}
-				const includeTransactions = getBlockByHashRequest.params[1] ?? false
-				const result = blockToJsonRpcBlock(block, includeTransactions)
-				return {
-					method: getBlockByHashRequest.method,
-					result,
-					jsonrpc: '2.0',
-					...(getBlockByHashRequest.id ? { id: getBlockByHashRequest.id } : {}),
-				}
-			}
-			case 'eth_getBlockTransactionCountByHash': {
-				const getBlockByHashRequest =
-					/** @type {import('./eth/EthJsonRpcRequest.js').EthGetBlockTransactionCountByHashJsonRpcRequest}*/
-					(request)
-				const vm = await client.getVm()
-				const block = await vm.blockchain.getBlock(hexToBytes(request.params[0]))
-				const result = block.transactions.length
-				return {
-					method: getBlockByHashRequest.method,
-					result: numberToHex(result),
-					jsonrpc: '2.0',
-					...(getBlockByHashRequest.id ? { id: getBlockByHashRequest.id } : {}),
-				}
-			}
-			case 'eth_getBlockTransactionCountByNumber': {
-				const getBlockByHashRequest =
-					/** @type {import('./eth/EthJsonRpcRequest.js').EthGetBlockTransactionCountByNumberJsonRpcRequest}*/
-					(request)
-				const vm = await client.getVm()
-				const blockTagOrNumber = getBlockByHashRequest.params[0]
-				const block = await (() => {
-					if (blockTagOrNumber.startsWith('0x')) {
-						return vm.blockchain.getBlock(hexToBigInt(/** @type {import('@tevm/utils').Hex}*/ (blockTagOrNumber)))
-					}
-					return vm.blockchain.blocksByTag.get(/** @type {import('@tevm/utils').BlockTag}*/ (blockTagOrNumber))
-				})()
-				if (!block) {
-					return {
-						...(request.id ? { id: request.id } : {}),
-						method: request.method,
-						jsonrpc: request.jsonrpc,
-						error: {
-							code: -32602,
-							message: `Invalid block tag ${blockTagOrNumber}`,
-						},
-					}
-				}
-				const result = block.transactions.length
-				return {
-					method: getBlockByHashRequest.method,
-					result: numberToHex(result),
-					jsonrpc: '2.0',
-					...(getBlockByHashRequest.id ? { id: getBlockByHashRequest.id } : {}),
-				}
-			}
-			case 'eth_getTransactionByHash': {
-				const getTransactionByHashRequest =
-					/** @type {import('./eth/EthJsonRpcRequest.js').EthGetTransactionByHashJsonRpcRequest}*/
-					(request)
-				const vm = await client.getVm()
-				const receiptsManager = await client.getReceiptsManager()
-				const receipt = await receiptsManager.getReceiptByTxHash(hexToBytes(request.params[0]))
-				if (!receipt && client.forkTransport) {
-					const fetcher = createJsonRpcFetcher(client.forkTransport)
-					return fetcher.request({
-						jsonrpc: '2.0',
-						id: request.id ?? 1,
-						method: 'eth_getTransactionByHash',
-						params: [request.params[0]],
-					})
-				}
-				if (!receipt) {
-					return {
-						...(request.id ? { id: request.id } : {}),
-						method: request.method,
-						jsonrpc: request.jsonrpc,
-						error: {
-							code: -32602,
-							message: 'Transaction not found',
-						},
-					}
-				}
-				const [_receipt, blockHash, txIndex] = receipt
-				const block = await vm.blockchain.getBlock(blockHash)
-				const tx = block.transactions[txIndex]
-				if (!tx) {
-					return {
-						...(request.id ? { id: request.id } : {}),
-						method: request.method,
-						jsonrpc: request.jsonrpc,
-						error: {
-							code: -32602,
-							message: 'Transaction not found',
-						},
-					}
-				}
-				return {
-					method: getTransactionByHashRequest.method,
-					result: txToJsonRpcTx(tx, block, txIndex),
-					jsonrpc: '2.0',
-					...(getTransactionByHashRequest.id ? { id: getTransactionByHashRequest.id } : {}),
-				}
-			}
-			case 'eth_getTransactionByBlockHashAndIndex': {
-				const getTransactionByBlockHashAndIndexRequest =
-					/** @type {import('./eth/EthJsonRpcRequest.js').EthGetTransactionByBlockHashAndIndexJsonRpcRequest}*/
-					(request)
-				const vm = await client.getVm()
-				const block = await vm.blockchain.getBlock(hexToBytes(request.params[0]))
-				const txIndex = hexToNumber(request.params[1])
-				const tx = block.transactions[txIndex]
-				if (!tx) {
-					return {
-						...(request.id ? { id: request.id } : {}),
-						method: request.method,
-						jsonrpc: request.jsonrpc,
-						error: {
-							code: -32602,
-							message: 'Transaction not found',
-						},
-					}
-				}
-				return {
-					method: getTransactionByBlockHashAndIndexRequest.method,
-					result: txToJsonRpcTx(tx, block, txIndex),
-					jsonrpc: '2.0',
-					...(getTransactionByBlockHashAndIndexRequest.id ? { id: getTransactionByBlockHashAndIndexRequest.id } : {}),
-				}
-			}
-			case 'eth_getTransactionByBlockNumberAndIndex': {
-				const getTransactionByBlockNumberAndIndexRequest =
-					/** @type {import('./eth/EthJsonRpcRequest.js').EthGetTransactionByBlockNumberAndIndexJsonRpcRequest}*/
-					(request)
-				const vm = await client.getVm()
-				const blockTagOrNumber = getTransactionByBlockNumberAndIndexRequest.params[0]
-				const block = await (() => {
-					if (blockTagOrNumber.startsWith('0x')) {
-						return vm.blockchain.getBlock(hexToBigInt(/** @type {import('@tevm/utils').Hex}*/ (blockTagOrNumber)))
-					}
-					return vm.blockchain.blocksByTag.get(/** @type {import('@tevm/utils').BlockTag}*/ (blockTagOrNumber))
-				})()
-				if (!block) {
-					return {
-						...(request.id ? { id: request.id } : {}),
-						method: request.method,
-						jsonrpc: request.jsonrpc,
-						error: {
-							code: -32602,
-							message: `Invalid block tag ${blockTagOrNumber}`,
-						},
-					}
-				}
-				const txIndex = hexToNumber(getTransactionByBlockNumberAndIndexRequest.params[1])
-				const tx = block.transactions[txIndex]
-				if (!tx) {
-					return {
-						...(request.id ? { id: request.id } : {}),
-						method: request.method,
-						jsonrpc: request.jsonrpc,
-						error: {
-							code: -32602,
-							message: 'Transaction not found',
-						},
-					}
-				}
-				return {
-					method: getTransactionByBlockNumberAndIndexRequest.method,
-					result: txToJsonRpcTx(tx, block, txIndex),
-					jsonrpc: '2.0',
-					...(getTransactionByBlockNumberAndIndexRequest.id
-						? { id: getTransactionByBlockNumberAndIndexRequest.id }
-						: {}),
-				}
-			}
-			// TODO add this to typescript type
-			// TODO this doesn't work yet see https://github.com/ethereumjs/ethereumjs-monorepo/pull/3436/files
-			// Getting some error with toBigInt internal to the block failing
-			case /** @type any*/ ('eth_blobBaseFee'): {
-				const vm = await client.getVm()
-				const headBlock = await vm.blockchain.getCanonicalHeadBlock()
-				return numberToHex(headBlock.header.calcNextBlobGasPrice())
-			}
-			case 'anvil_reset': {
-				const vm = await client.getVm()
-				vm.blockchain.blocksByTag.set(
-					'latest',
-					vm.blockchain.blocksByTag.get('forked') ?? vm.blockchain.blocksByTag.get('latest'),
-				)
-				Array.from(vm.blockchain.blocks.values()).forEach((block) => {
-					if (!block) return
-					vm.blockchain.delBlock(block.hash())
-				})
-				const stateManager = vm.stateManager.shallowCopy()
-				vm.stateManager = /** @type any*/ (stateManager)
-				vm.evm.stateManager = /** @type any*/ (stateManager)
-				return {
-					method: request.method,
-					jsonrpc: '2.0',
-					...(request.id ? { id: request.id } : {}),
-				}
-			}
-			case 'anvil_setStorageAt':
-			case /** @type any*/ ('ganache_setStorageAt'):
-			case /** @type any*/ ('hardhat_setStorageAt'): {
-				const anvilSetStorageAtRequest =
-					/** @type {import('./anvil/AnvilJsonRpcRequest.js').AnvilSetStorageAtJsonRpcRequest}*/
-					(request)
-				const result = await setAccountProcedure(client)({
-					method: 'tevm_setAccount',
-					...(anvilSetStorageAtRequest.id ? { id: anvilSetStorageAtRequest.id } : {}),
-					jsonrpc: '2.0',
-					params: [
-						{
-							address: anvilSetStorageAtRequest.params[0],
-							stateDiff: {
-								[anvilSetStorageAtRequest.params[1]]: anvilSetStorageAtRequest.params[2],
-							},
-						},
-					],
-				})
-				return {
-					...result,
-					method: anvilSetStorageAtRequest.method,
-				}
-			}
-			case 'anvil_dropTransaction': {
-				const anvilDropTransactionRequest =
-					/** @type {import('./anvil/AnvilJsonRpcRequest.js').AnvilDropTransactionJsonRpcRequest}*/
-					(request)
-				const txHash = anvilDropTransactionRequest.params[0].transactionHash
-				const txPool = await client.getTxPool()
-				if (txPool.getByHash([hexToBytes(txHash)]).length > 0) {
-					txPool.removeByHash(txHash)
-				} else {
-					throw new Error(
-						'Only tx in the txpool are allowed to be dropped. Dropping transactions that have already been mined is not yet supported',
-					)
-				}
-				return {
-					method: anvilDropTransactionRequest.method,
-					jsonrpc: '2.0',
-					...(anvilDropTransactionRequest.id ? { id: anvilDropTransactionRequest.id } : {}),
-				}
-			}
-			case 'anvil_dumpState':
-				return {
-					...(await dumpStateProcedure(client)({
-						...(request.id ? { id: request.id } : {}),
-						jsonrpc: '2.0',
-						method: 'tevm_dumpState',
-					})),
-					method: request.method,
-				}
-			case 'eth_protocolVersion': {
-				return {
-					result: packageJsonVersion,
-					jsonrpc: '2.0',
-					method: 'eth_protocolVersion',
-					...(request.id ? { id: request.id } : {}),
-				}
-			}
-			case 'anvil_loadState': {
-				const loadStateRequest = /** @type {import('./anvil/AnvilJsonRpcRequest.js').AnvilLoadStateJsonRpcRequest}*/ (
-					request
-				)
-
-				const vm = await client.getVm()
-
-				return Promise.all(
-					Object.entries(loadStateRequest.params[0].state).map(([address, rlpEncodedAccount]) => {
-						return vm.stateManager.putAccount(
-							EthjsAddress.fromString(address),
-							EthjsAccount.fromRlpSerializedAccount(hexToBytes(rlpEncodedAccount)),
-						)
-					}),
-				)
-					.then(() => {
-						/**
-						 * @type {import('./anvil/AnvilJsonRpcResponse.js').AnvilLoadStateJsonRpcResponse}
-						 */
-						return {
-							jsonrpc: '2.0',
-							method: loadStateRequest.method,
-							...(loadStateRequest.id ? { id: loadStateRequest.id } : {}),
-						}
-					})
-					.catch((e) => {
-						return {
-							jsonrpc: '2.0',
-							method: loadStateRequest.method,
-							...(loadStateRequest.id ? { id: loadStateRequest.id } : {}),
-							error: {
-								code: -32602,
-								message: e.message,
-							},
-						}
-					})
-			}
-			case 'debug_traceTransaction': {
-				const debugTraceTransactionRequest =
-					/** @type {import('./debug/DebugJsonRpcRequest.js').DebugTraceTransactionJsonRpcRequest}*/
-					(request)
-				const { tracer, timeout, tracerConfig, transactionHash } = request.params[0]
-				if (timeout !== undefined) {
-					client.logger.warn('Warning: timeout is currently respected param of debug_traceTransaction')
-				}
-				const transactionByHashResponse = await requestProcedure(client)({
-					method: 'eth_getTransactionByHash',
-					params: [transactionHash],
-					jsonrpc: '2.0',
-					id: 1,
-				})
-				if ('error' in transactionByHashResponse) {
-					return {
-						...transactionByHashResponse,
-						method: debugTraceTransactionRequest.method,
-					}
-				}
-				const vm = await client.getVm()
-				const block = await vm.blockchain.getBlock(hexToBytes(transactionByHashResponse.result.blockHash))
-				const parentBlock = await vm.blockchain.getBlock(block.header.parentHash)
-				transactionByHashResponse.result.transactionIndex
-				const previousTx = block.transactions.filter(
-					(_, i) => i < hexToNumber(transactionByHashResponse.result.transactionIndex),
-				)
-				const hasStateRoot = vm.stateManager.hasStateRoot(parentBlock.header.stateRoot)
-				if (!hasStateRoot && client.forkTransport) {
-					await forkAndCacheBlock(client, parentBlock)
-				} else {
-					return {
-						jsonrpc: '2.0',
-						method: debugTraceTransactionRequest.method,
-						id: debugTraceTransactionRequest.id,
-						error: {
-							code: -32602,
-							message: 'Parent block not found',
-						},
-					}
-				}
-				const vmClone = await vm.deepCopy()
-
-				// execute all transactions before the current one committing to the state
-				for (const tx of previousTx) {
-					runTx(vmClone)({
-						block: parentBlock,
-						skipNonce: true,
-						skipBalance: true,
-						skipHardForkValidation: true,
-						skipBlockGasLimitValidation: true,
-						tx: await TransactionFactory.fromRPC(tx, {
-							freeze: false,
-							common: vmClone.common.ethjsCommon,
-							allowUnlimitedInitCodeSize: true,
-						}),
-					})
-				}
-
-				// now execute an debug_traceCall
-				const traceResult = await traceCallHandler(client)({
-					tracer,
-					...(transactionByHashResponse.result.to !== undefined ? { to: transactionByHashResponse.result.to } : {}),
-					...(transactionByHashResponse.result.from !== undefined
-						? { from: transactionByHashResponse.result.from }
-						: {}),
-					...(transactionByHashResponse.result.gas !== undefined
-						? { gas: hexToBigInt(transactionByHashResponse.result.gas) }
-						: {}),
-					...(transactionByHashResponse.result.gasPrice !== undefined
-						? { gasPrice: hexToBigInt(transactionByHashResponse.result.gasPrice) }
-						: {}),
-					...(transactionByHashResponse.result.value !== undefined
-						? { value: hexToBigInt(transactionByHashResponse.result.value) }
-						: {}),
-					...(transactionByHashResponse.result.data !== undefined
-						? { data: transactionByHashResponse.result.data }
-						: {}),
-					...(transactionByHashResponse.result.blockHash !== undefined
-						? { blockTag: transactionByHashResponse.result.blockHash }
-						: {}),
-					...(timeout !== undefined ? { timeout } : {}),
-					...(tracerConfig !== undefined ? { tracerConfig } : {}),
-				})
-				return {
-					method: debugTraceTransactionRequest.method,
-					result: {
-						gas: numberToHex(traceResult.gas),
-						failed: traceResult.failed,
-						returnValue: traceResult.returnValue,
-						structLogs: traceResult.structLogs.map((log) => {
-							return {
-								gas: numberToHex(log.gas),
-								gasCost: numberToHex(log.gasCost),
-								op: log.op,
-								pc: log.pc,
-								stack: log.stack,
-								depth: log.depth,
-							}
-						}),
-					},
-					jsonrpc: '2.0',
-					...(debugTraceTransactionRequest.id ? { id: debugTraceTransactionRequest.id } : {}),
-				}
-			}
-			case 'anvil_impersonateAccount': {
-				const impersonateAccountRequest =
-					/** @type {import('./anvil/AnvilJsonRpcRequest.js').AnvilImpersonateAccountJsonRpcRequest}*/
-					(request)
-				try {
-					client.setImpersonatedAccount(getAddress(impersonateAccountRequest.params[0]))
-					return {
-						jsonrpc: '2.0',
-						method: impersonateAccountRequest.method,
-						id: impersonateAccountRequest.id,
-						result: true,
-					}
-				} catch (e) {
-					return {
-						jsonrpc: '2.0',
-						method: impersonateAccountRequest.method,
-						id: impersonateAccountRequest.id,
-						error: {
-							code: -32602,
-							message: /** @type {Error}*/ (e).message,
-						},
-					}
-				}
-			}
-			case 'anvil_stopImpersonatingAccount': {
-				client.setImpersonatedAccount(undefined)
-				return {
-					jsonrpc: '2.0',
-					method: request.method,
-					id: request.id,
-					result: true,
-				}
-			}
-			case 'eth_getTransactionCount': {
-				const [address, tag] = request.params
-
-				const block = await (async () => {
-					const vm = await client.getVm()
-					if (tag.startsWith('0x') && tag.length === 66) {
-						return vm.blockchain.getBlock(hexToBytes(/** @type {import('@tevm/utils').Hex}*/ (tag)))
-					}
-					if (tag.startsWith('0x')) {
-						return vm.blockchain.getBlock(hexToBigInt(/** @type {import('@tevm/utils').Hex}*/ (tag)))
-					}
-					if (tag === 'pending') {
-						// if pending just get latest we will get pending seperately
-						return vm.blockchain.blocksByTag.get('latest')
-					}
-					if (tag === 'latest' || tag === 'safe' || tag === 'earliest' || tag === 'finalized') {
-						return vm.blockchain.blocksByTag.get(tag)
-					}
-					return undefined
-				})()
-				if (!block) {
-					return {
-						...(request.id ? { id: request.id } : {}),
-						method: request.method,
-						error: {
-							code: -32602,
-							message: `Invalid block tag ${tag}`,
-						},
-					}
-				}
-
-				const pendingCount =
-					tag === 'pending'
-						? await (async () => {
-								const txPool = await client.getTxPool()
-								const pendingTx = await txPool.getBySenderAddress(EthjsAddress.fromString(address))
-								return BigInt(pendingTx.length)
-							})()
-						: 0n
-
-				const includedCount = await (async () => {
-					const vm = await client.getVm()
-					// TODO we can optimize this by not deep copying once we are more confident it's safe
-					const root = vm.stateManager._baseState.stateRoots.get(bytesToHex(block.header.stateRoot))
-					if (!root) {
-						// todo we might want to throw an error hre
-						return 0n
-					}
-					return root[getAddress(address)]?.nonce ?? 0n
-				})()
-
-				return {
-					...(request.id ? { id: request.id } : {}),
-					method: request.method,
-					jsonrpc: request.jsonrpc,
-					result: numberToHex(pendingCount + includedCount),
-				}
-			}
-			case 'eth_newFilter': {
-				const newFilterRequest = /** @type {import('./eth/EthJsonRpcRequest.js').EthNewFilterJsonRpcRequest}*/ (request)
-
-				const { topics, address, toBlock = 'latest', fromBlock = 'latest' } = newFilterRequest.params[0]
-				const id = generateRandomId()
-				const vm = await client.getVm()
-				/**
-				 * @param {typeof toBlock} tag
-				 */
-				const getBlock = async (tag) => {
-					const parsedTag = parseBlockTag(tag)
-					if (
-						parsedTag === 'safe' ||
-						parsedTag === 'latest' ||
-						parsedTag === 'finalized' ||
-						parsedTag === 'earliest' ||
-						parsedTag === 'pending' ||
-						parsedTag === /** @type any*/ ('forked')
-					) {
-						return vm.blockchain.blocksByTag.get(parsedTag)
-					}
-					if (typeof parsedTag === 'string') {
-						return vm.blockchain.getBlock(hexToBytes(parsedTag))
-					}
-					return vm.blockchain.getBlock(parsedTag)
-				}
-				const _toBlock = await getBlock(toBlock)
-				if (!_toBlock) {
-					return {
-						...(request.id ? { id: request.id } : {}),
-						method: request.method,
-						jsonrpc: request.jsonrpc,
-						error: {
-							code: -32602,
-							message: `Invalid block tag ${toBlock}`,
-						},
-					}
-				}
-				const _fromBlock = await getBlock(fromBlock)
-				if (!_fromBlock) {
-					return {
-						...(request.id ? { id: request.id } : {}),
-						method: request.method,
-						jsonrpc: request.jsonrpc,
-						error: {
-							code: -32602,
-							message: `Invalid block tag ${fromBlock}`,
-						},
-					}
-				}
-
-				/**
-				 * @param {import('@tevm/base-client').Filter['logs'][number]} log
-				 */
-				const listener = (log) => {
-					const filter = client.getFilters().get(id)
-					if (!filter) {
-						return
-					}
-					filter.logs.push(log)
-				}
-				client.on('newLog', listener)
-				// populate with past blocks
-				const receiptsManager = await client.getReceiptsManager()
-				const pastLogs = await receiptsManager.getLogs(
-					_fromBlock,
-					_toBlock,
-					address !== undefined ? [EthjsAddress.fromString(address).bytes] : [],
-					topics?.map((topic) => hexToBytes(topic)),
-				)
-				client.setFilter({
-					id,
-					type: 'Log',
-					created: Date.now(),
-					logs: pastLogs.map((log) => {
-						const [address, topics, data] = log.log
-						return {
-							topics: /** @type {[import('@tevm/utils').Hex, ...Array<import('@tevm/utils').Hex>]}*/ (
-								topics.map((topic) => bytesToHex(topic))
-							),
-							address: bytesToHex(address),
-							data: bytesToHex(data),
-							blockNumber: log.block.header.number,
-							transactionHash: bytesToHex(log.tx.hash()),
-							removed: false,
-							logIndex: log.logIndex,
-							blockHash: bytesToHex(log.block.hash()),
-							transactionIndex: log.txIndex,
-						}
-					}),
-					tx: [],
-					blocks: [],
-					logsCriteria: {
-						topics,
-						address,
-						toBlock: toBlock,
-						fromBlock: fromBlock,
-					},
-					installed: {},
-					err: undefined,
-					registeredListeners: [listener],
-				})
-				return {
-					...(request.id ? { id: request.id } : {}),
-					method: request.method,
-					jsonrpc: request.jsonrpc,
-					result: id,
-				}
-			}
-			case 'eth_getFilterLogs': {
-				const ethGetFilterLogsRequest =
-					/** @type {import('./eth/EthJsonRpcRequest.js').EthGetFilterLogsJsonRpcRequest}*/
-					(request)
-				const filter = client.getFilters().get(ethGetFilterLogsRequest.params[0])
-				if (!filter) {
-					return {
-						...(ethGetFilterLogsRequest.id ? { id: ethGetFilterLogsRequest.id } : {}),
-						method: ethGetFilterLogsRequest.method,
-						jsonrpc: ethGetFilterLogsRequest.jsonrpc,
-						error: {
-							code: -32602,
-							message: 'Filter not found',
-						},
-					}
-				}
-				try {
-					const ethGetLogsResult = await ethGetLogsHandler(client)({
-						filterParams: {
-							fromBlock: filter.logsCriteria.fromBlock,
-							toBlock: filter.logsCriteria.toBlock,
-							address: filter.logsCriteria.address,
-							topics: filter.logsCriteria.topics,
-						},
-					})
-					// TODO move this to a shared util method since it's copy pasted from the eth_getLogs procedure
-					const jsonRpcResult = ethGetLogsResult.map((log) => ({
-						address: log.address,
-						topics: log.topics,
-						data: log.data,
-						blockNumber: numberToHex(log.blockNumber),
-						transactionHash: log.transactionHash,
-						transactionIndex: numberToHex(log.transactionIndex),
-						blockHash: log.blockHash,
-						logIndex: numberToHex(log.logIndex),
-						removed: log.removed,
-					}))
-					return {
-						...(request.id ? { id: request.id } : {}),
-						method: request.method,
-						jsonrpc: request.jsonrpc,
-						result: jsonRpcResult,
-					}
-				} catch (e) {
-					return {
-						...(ethGetFilterLogsRequest.id ? { id: ethGetFilterLogsRequest.id } : {}),
-						method: ethGetFilterLogsRequest.method,
-						jsonrpc: ethGetFilterLogsRequest.jsonrpc,
-						error: {
-							code: -32601,
-							message: /** @type {Error}*/ (e).message,
-						},
-					}
-				}
-			}
-			case 'eth_newBlockFilter': {
-				const newBlockFilterRequest =
-					/** @type {import('./eth/EthJsonRpcRequest.js').EthNewBlockFilterJsonRpcRequest}*/
-					(request)
-				const id = generateRandomId()
-				/**
-				 * @param {import('@tevm/block').Block} block
-				 */
-				const listener = (block) => {
-					const filter = client.getFilters().get(id)
-					if (!filter) {
-						return
-					}
-					filter.blocks.push(block)
-				}
-				client.setFilter({
-					id,
-					type: 'Block',
-					created: Date.now(),
-					logs: [],
-					tx: [],
-					blocks: [],
-					installed: {},
-					err: undefined,
-					registeredListeners: [listener],
-				})
-				return {
-					...(newBlockFilterRequest.id ? { id: newBlockFilterRequest.id } : {}),
-					method: newBlockFilterRequest.method,
-					jsonrpc: newBlockFilterRequest.jsonrpc,
-					result: id,
-				}
-			}
-			case 'eth_uninstallFilter': {
-				const uninstallFilterRequest =
-					/** @type {import('./eth/EthJsonRpcRequest.js').EthUninstallFilterJsonRpcRequest}*/
-					(request)
-				const [filterId] = uninstallFilterRequest.params
-				const filter = client.getFilters().get(filterId)
-				if (!filter) {
-					return {
-						...(uninstallFilterRequest.id ? { id: uninstallFilterRequest.id } : {}),
-						method: uninstallFilterRequest.method,
-						jsonrpc: uninstallFilterRequest.jsonrpc,
-						result: false,
-					}
-				}
-
-				const [listener] = filter.registeredListeners
-				if (filter.type === 'Log' && listener) {
-					client.removeListener('newLog', listener)
-				} else if (filter.type === 'Block' && listener) {
-					client.removeListener('newBlock', listener)
-				} else if (filter.type === 'PendingTransaction' && listener) {
-					client.removeListener('newPendingTransaction', listener)
-				}
-				client.removeFilter(filterId)
-				return {
-					...(request.id ? { id: request.id } : {}),
-					method: request.method,
-					jsonrpc: request.jsonrpc,
-					result: true,
-				}
-			}
-			case 'eth_getFilterChanges': {
-				const getFilterChangesRequest =
-					/** @type {import('./eth/EthJsonRpcRequest.js').EthGetFilterChangesJsonRpcRequest}*/
-					(request)
-				const [id] = getFilterChangesRequest.params
-				const filter = client.getFilters().get(id)
-				if (!filter) {
-					return {
-						...(request.id ? { id: request.id } : {}),
-						method: request.method,
-						jsonrpc: request.jsonrpc,
-						error: {
-							code: -32601,
-							message: 'Method not implemented yet',
-						},
-					}
-				}
-				switch (filter.type) {
-					case 'Log': {
-						const { logs } = filter
-						/**
-						 * @type {import('./eth/EthJsonRpcResponse.js').EthGetFilterChangesJsonRpcResponse}
-						 */
-						const response = {
-							...(request.id ? { id: request.id } : {}),
-							method: request.method,
-							jsonrpc: request.jsonrpc,
-							result: logs.map((log) => ({
-								address: log.address,
-								topics: log.topics,
-								data: log.data,
-								blockNumber: numberToHex(log.blockNumber),
-								transactionHash: log.transactionHash,
-								transactionIndex: numberToHex(log.transactionIndex),
-								blockHash: log.blockHash,
-								logIndex: numberToHex(log.logIndex),
-								removed: log.removed,
-							})),
-						}
-						filter.logs = []
-						return response
-					}
-					case 'Block': {
-						const { blocks } = filter
-						/**
-						 * @type {import('./eth/EthJsonRpcResponse.js').EthGetFilterChangesJsonRpcResponse}
-						 */
-						const response = {
-							...(request.id ? { id: request.id } : {}),
-							// TODO fix this type
-							result: /** @type {any} */ (blocks.map((block) => numberToHex(block.header.number))),
-							method: request.method,
-							jsonrpc: request.jsonrpc,
-						}
-						filter.blocks = []
-						return response
-					}
-					case 'PendingTransaction': {
-						const { tx } = filter
-						/**
-						 * @type {import('./eth/EthJsonRpcResponse.js').EthGetFilterChangesJsonRpcResponse}
-						 */
-						const response = {
-							...(request.id ? { id: request.id } : {}),
-							// TODO fix this type
-							result: /** @type {any} */ (tx.map((tx) => bytesToHex(tx.hash()))),
-							method: request.method,
-							jsonrpc: request.jsonrpc,
-						}
-						filter.tx = []
-						return response
-					}
-					default: {
-						throw new Error(
-							'InternalError: Unknown filter type. This indicates a bug in tevm or potentially a typo in filter type if manually added',
-						)
-					}
-				}
-			}
-			case 'eth_newPendingTransactionFilter': {
-				const newPendingTransactionFilterRequest =
-					/** @type {import('./eth/EthJsonRpcRequest.js').EthNewPendingTransactionFilterJsonRpcRequest}*/
-					(request)
-				const id = generateRandomId()
-				/**
-				 * @param {import('@tevm/tx').TypedTransaction} tx
-				 */
-				const listener = (tx) => {
-					const filter = client.getFilters().get(id)
-					if (!filter) {
-						return
-					}
-					filter.tx.push(tx)
-				}
-				client.on('newPendingTransaction', listener)
-				client.setFilter({
-					id,
-					type: 'PendingTransaction',
-					created: Date.now(),
-					logs: [],
-					tx: [],
-					blocks: [],
-					installed: {},
-					err: undefined,
-					registeredListeners: [listener],
-				})
-				return {
-					...(newPendingTransactionFilterRequest.id ? { id: newPendingTransactionFilterRequest.id } : {}),
-					method: newPendingTransactionFilterRequest.method,
-					jsonrpc: newPendingTransactionFilterRequest.jsonrpc,
-					result: id,
-				}
-			}
-
-			default: {
-				const err = new MethodNotFoundError(`UnsupportedMethodError: Unknown method ${/**@type any*/ (request).method}`)
-				return /** @type {any}*/ ({
-					id: /** @type any*/ (request).id ?? null,
-					method: /** @type any*/ (request).method,
-					jsonrpc: '2.0',
-					error: {
-						code: 4200,
-						message: err.message,
-					},
-				})
-			}
+		if (!(request.method in allHandlers)) {
+			const err = new MethodNotFoundError(`UnsupportedMethodError: Unknown method ${/**@type any*/ (request).method}`)
+			return /** @type {any}*/ ({
+				id: /** @type any*/ (request).id ?? null,
+				method: /** @type any*/ (request).method,
+				jsonrpc: '2.0',
+				error: {
+					code: err.code,
+					message: err.message,
+				},
+			})
 		}
+		return allHandlers[/** @type {keyof typeof allHandlers}*/ (request.method)](client)(request)
 	}
 }
